LayoutTests/animations/trigger-parsing-expected.txt

CONSOLE MESSAGE: line 37: TypeError: null is not an object (evaluating 'triggerRule[i]')
Test the parsing of the -webkit-animation-trigger property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Valid property tests


Auto : auto
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-animation-trigger') should be auto. Was .
FAIL triggerRule.cssText should be auto. Threw exception TypeError: null is not an object (evaluating 'triggerRule.cssText')

Container Scroll : container-scroll(200px)
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-animation-trigger') should be container-scroll(200px). Was .
FAIL triggerRule.cssText should be container-scroll(200px). Threw exception TypeError: null is not an object (evaluating 'triggerRule.cssText')

Container Scroll : container-scroll(2em)
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-animation-trigger') should be container-scroll(2em). Was .
FAIL triggerRule.cssText should be container-scroll(2em). Threw exception TypeError: null is not an object (evaluating 'triggerRule.cssText')

Container Scroll : container-scroll(30vw)
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-animation-trigger') should be container-scroll(30vw). Was .
FAIL triggerRule.cssText should be container-scroll(30vw). Threw exception TypeError: null is not an object (evaluating 'triggerRule.cssText')

Container Scroll with end value : container-scroll(200px, 300px)
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-animation-trigger') should be container-scroll(200px, 300px). Was .
FAIL triggerRule.cssText should be container-scroll(200px, 300px). Threw exception TypeError: null is not an object (evaluating 'triggerRule.cssText')

Two scroll values : container-scroll(1px), container-scroll(2px)
PASS cssRule.type is 1
FAIL declaration.length should be 1. Was 0.
FAIL declaration.getPropertyValue('-webkit-animation-trigger') should be container-scroll(1px), container-scroll(2px). Was .
FAIL triggerRule.length should be 2. Threw exception TypeError: null is not an object (evaluating 'triggerRule.length')
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
